backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cffi==1.17.0
contextlib2==21.6.0
cryptography==43.0.0
Django==4.2
django-cors-headers==4.4.0
django-extensions==3.2.3
django-ninja==1.2.2
django-ninja-extra==0.21.3
django-ninja-jwt==5.3.2
gunicorn==22.0.0
injector==0.22.0
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.1
typing_extensions==4.12.2
